Refactored translators
