+++ xnmt/models/__init__.py
@@ -1,0 +1,1 @@
+from xnmt.models.base import ConditionedModel, GeneratorModel
