+++ xnmt/transducers/__init__.py
@@ -1,0 +1,2 @@
+from xnmt.transducers.recurrent import BiLSTMSeqTransducer
+
